Move git code to it's own folder
